+++ opencascade-hs/cpp/hs_types.h
@@ -75,6 +75,7 @@ typedef void Message_ProgressRange;
 typedef void RWGltf_CafWriter;
 typedef void RWGltf_CafReader;
 typedef void RWObj_CafWriter;
+typedef void RWObj_CafReader;
 typedef void RWMesh_CafReader;
 typedef void TColStd_IndexedDataMapOfStringString;
 typedef void TDF_Label;